@@ -1,10 +1,10 @@
 name: color_picker_example
 description: Two examples of how to use the FlexColorPicker package.
-version: 3.4.1
+version: 3.5.0-dev.1
 publish_to: 'none'
 environment:
   sdk: '>=2.19.0 <4.0.0'
-  flutter: '>=3.19.0'
+  flutter: '>=3.22.0-0.3.pre'
 
 dependencies:
   # https://pub.dev/packages/flex_color_picker
@@ -24,11 +24,11 @@ dependencies:
   hive: ^2.2.3
 
   # https://pub.dev/packages/path_provider:
-  path_provider: ^2.1.2
+  path_provider: ^2.1.3
 
   # https://pub.dev/packages/url_launcher
   # Used for launching a WEB URL (by Google flutter.dev).
-  url_launcher: ^6.2.5
+  url_launcher: ^6.2.6
 
 dev_dependencies:
   flutter_test:
