@@ -730,7 +730,7 @@ class FlexAppBar {
     // flag to use it is true OR if the flag to use it on dark, black or white
     // app bars is true and we have dark, black or white color in the app bar.
     final Color schemeSurface = theme.colorScheme.surface;
-    final Color schemeBackground = theme.colorScheme.background;
+    final Color schemeBackground = theme.colorScheme.surface;
     final bool effectiveBottomBorder = hasBorder ||
         (hasBorderOnSurface &&
             (appBarColor == Colors.black ||
